+++ android/src/main/java/com/pusherwebsocketreactnative/PusherWebsocketReactNativeModule.kt
@@ -65,9 +65,9 @@ class PusherWebsocketReactNativeModule(reactContext: ReactApplicationContext) :
         if (arguments.hasKey("useTLS")) options.isUseTLS =
           arguments.getBoolean("useTLS")
         if (arguments.hasKey("activityTimeout")) options.activityTimeout =
-          arguments.getInt("activityTimeout") as Long
+          arguments.getInt("activityTimeout").toLong()
         if (arguments.hasKey("pongTimeout")) options.pongTimeout =
-          arguments.getInt("pongTimeout") as Long
+          arguments.getInt("pongTimeout").toLong()
         if (arguments.hasKey("maxReconnectionAttempts")) options.maxReconnectionAttempts =
           arguments.getInt("maxReconnectionAttempts")
         if (arguments.hasKey("maxReconnectGapInSeconds")) options.maxReconnectGapInSeconds =
@@ -160,8 +160,7 @@ class PusherWebsocketReactNativeModule(reactContext: ReactApplicationContext) :
     authorizerMutex[key]!!.acquire()
     val authParams = authorizerResult.remove(key)!!
     val gson = Gson()
-    val json = gson.toJson(authParams.toHashMap())
-    return json
+    return gson.toJson(authParams.toHashMap())
   }
 
   @ReactMethod
@@ -198,12 +197,23 @@ class PusherWebsocketReactNativeModule(reactContext: ReactApplicationContext) :
 
   override fun onEvent(event: PusherEvent) {
     // Log.i(TAG, "Received event with data: $event")
+    // The java sdk transforms some events from pusher_internal
+    // to pusher:... events, we translate them back.
+    val finalEvent = if (event.eventName === "pusher:subscription_count") {
+      PusherEvent(
+        "pusher_internal:subscription_count",
+        event.channelName,
+        event.userId,
+        event.data)
+    } else {
+      event
+    }
     emitEvent(
       "onEvent", mapOf(
-        "channelName" to event.channelName,
-        "eventName" to event.eventName,
-        "userId" to event.userId,
-        "data" to event.data
+        "channelName" to finalEvent.channelName,
+        "eventName" to finalEvent.eventName,
+        "userId" to finalEvent.userId,
+        "data" to finalEvent.data
       )
     )
   }